% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange.R
\name{arrange.RPolarsDataFrame}
\alias{arrange.RPolarsDataFrame}
\title{Order rows using column values}
\usage{
\method{arrange}{RPolarsDataFrame}(.data, ..., .by_group = FALSE)
}
\arguments{
\item{.data}{A Polars Data/LazyFrame}

\item{...}{Variables, or functions of variables. Use \code{desc()} to sort a
variable in descending order.}

\item{.by_group}{If \code{TRUE}, will sort data within groups.}
}
\description{
Order rows using column values
}
\examples{
\dontshow{if (require("dplyr", quietly = TRUE) && require("tidyr", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
pl_test <- polars::pl$DataFrame(
  x1 = c("a", "a", "b", "a", "c"),
  x2 = c(2, 1, 5, 3, 1),
  value = sample(1:5)
)

arrange(pl_test, x1)
arrange(pl_test, x1, -x2)

# if the data is grouped, you need to specify `.by_group = TRUE` to sort by
# the groups first
pl_test |>
  group_by(x1) |>
  arrange(-x2, .by_group = TRUE)
\dontshow{\}) # examplesIf}
}
