% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_polars.R
\name{as_polars}
\alias{as_polars}
\title{Convert an R dataframe to a Polars Data/LazyFrame}
\usage{
as_polars(data, lazy = FALSE)
}
\arguments{
\item{data}{An R dataframe.}

\item{lazy}{Convert the data to lazy format.}
}
\description{
This operation is time- and memory-consuming. It should only be used for small
datasets. Use \code{polars} original functions to read files instead.
}
\examples{
mtcars |>
  as_polars()
}
