% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind.R
\name{bind_cols_polars}
\alias{bind_cols_polars}
\title{Append multiple Data/LazyFrames next to each other}
\usage{
bind_cols_polars(..., .name_repair = "unique")
}
\arguments{
\item{...}{Polars DataFrames or LazyFrames to combine. Each argument can
either be a Data/LazyFrame, or a list of Data/LazyFrames. Columns are matched
by name. All Data/LazyFrames must have the same number of rows and there
mustn't be duplicated column names.}

\item{.name_repair}{Can be \code{"unique"}, \code{"universal"}, \code{"check_unique"},
\code{"minimal"}. See \code{\link[vctrs:vec_as_names]{vctrs::vec_as_names()}} for the explanations for each value.}
}
\description{
Append multiple Data/LazyFrames next to each other
}
\examples{
\dontshow{if (require("dplyr", quietly = TRUE) && require("tidyr", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
p1 <- polars::pl$DataFrame(
  x = sample(letters, 20),
  y = sample(1:100, 20)
)
p2 <- polars::pl$DataFrame(
  z = sample(letters, 20),
  w = sample(1:100, 20)
)

bind_cols_polars(p1, p2)
bind_cols_polars(list(p1, p2))
\dontshow{\}) # examplesIf}
}
