% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind.R
\name{bind_rows_polars}
\alias{bind_rows_polars}
\title{Stack multiple Data/LazyFrames on top of each other}
\usage{
bind_rows_polars(..., .id = NULL)
}
\arguments{
\item{...}{Polars DataFrames or LazyFrames to combine. Each argument can
either be a Data/LazyFrame, or a list of Data/LazyFrames. Columns are matched
by name. All Data/LazyFrames must have the same number of columns with
identical names.}

\item{.id}{The name of an optional identifier column. Provide a string to
create an output column that identifies each input.}
}
\description{
Stack multiple Data/LazyFrames on top of each other
}
\examples{
\dontshow{if (require("dplyr", quietly = TRUE) && require("tidyr", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
p1 <- polars::pl$DataFrame(
  x = sample(letters, 20),
  y = sample(1:100, 20)
)
p2 <- polars::pl$DataFrame(
  x = sample(letters, 20),
  y = sample(1:100, 20)
)

bind_rows_polars(p1, p2)

# this is equivalent
bind_rows_polars(list(p1, p2))

# create an id colum
bind_rows_polars(p1, p2, .id = "id")
\dontshow{\}) # examplesIf}
}
