% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{count.DataFrame}
\alias{count.DataFrame}
\title{Count the observations in each group}
\usage{
\method{count}{DataFrame}(x, ..., sort = FALSE, name = "n")
}
\arguments{
\item{x}{A Polars Data/LazyFrame}

\item{...}{Any expression accepted by \code{dplyr::select()}: variable names,
column numbers, select helpers, etc.}

\item{sort}{If \code{TRUE}, will show the largest groups at the top.}

\item{name}{Name of the new column.}
}
\description{
Count the observations in each group
}
\examples{
\dontshow{if (require("dplyr", quietly = TRUE) && require("tidyr", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
test <- polars::pl$DataFrame(mtcars)
count(test, cyl)

count(test, cyl, am)

count(test, cyl, am, sort = TRUE, name = "count")

add_count(test, cyl, am, sort = TRUE, name = "count")
\dontshow{\}) # examplesIf}
}
