% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{count.DataFrame}
\alias{count.DataFrame}
\alias{add_count.DataFrame}
\title{Count the observations in each group}
\usage{
\method{count}{DataFrame}(.data, ..., sort = FALSE, name = "n")

\method{add_count}{DataFrame}(.data, ..., sort = FALSE, name = "n")
}
\arguments{
\item{.data}{A Polars Data/LazyFrame}

\item{sort}{If \code{TRUE}, will show the largest groups at the top.}

\item{name}{Name of the new column.}
}
\description{
Count the observations in each group
}
\examples{
test <- polars::pl$DataFrame(mtcars)
count(test, cyl)

count(test, cyl, am)

count(test, cyl, am, sort = TRUE, name = "count")

add_count(test, cyl, am, sort = TRUE, name = "count")
}
