% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.R
\name{cross_join.DataFrame}
\alias{cross_join.DataFrame}
\title{Cross join}
\usage{
\method{cross_join}{DataFrame}(x, y, suffix = c(".x", ".y"))
}
\description{
Cross joins match each row in \code{x} to every row in \code{y}, resulting in a dataset
with \code{nrow(x) * nrow(y)} rows.
}
\examples{
test <- polars::pl$DataFrame(
  origin = c("ALG", "FRA", "GER"),
  year = c(2020, 2020, 2021)
)

test2 <- polars::pl$DataFrame(
  destination = c("USA", "JPN", "BRA"),
  language = c("english", "japanese", "portuguese")
)

test

test2

cross_join(test, test2)
}
