% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.R
\name{cross_join.polars_data_frame}
\alias{cross_join.polars_data_frame}
\alias{cross_join.polars_lazy_frame}
\title{Cross join}
\usage{
\method{cross_join}{polars_data_frame}(x, y, ..., suffix = c(".x", ".y"))

\method{cross_join}{polars_lazy_frame}(x, y, ..., suffix = c(".x", ".y"))
}
\arguments{
\item{x, y}{Two Polars Data/LazyFrames}

\item{...}{Dots which should be empty.}

\item{suffix}{If there are non-joined duplicate variables in \code{x} and \code{y},
these suffixes will be added to the output to disambiguate them. Should be a
character vector of length 2.}
}
\description{
Cross joins match each row in \code{x} to every row in \code{y}, resulting in a dataset
with \code{nrow(x) * nrow(y)} rows.
}
\section{Unknown arguments}{


Arguments that are supported by the original implementation in the tidyverse
but are not listed above will throw a warning by default if they are
specified. To change this behavior to error instead, use
\code{options(tidypolars_unknown_args = "error")}.
}

\examples{
\dontshow{if (require("dplyr", quietly = TRUE) && require("tidyr", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
test <- polars::pl$DataFrame(
  origin = c("ALG", "FRA", "GER"),
  year = c(2020, 2020, 2021)
)

test2 <- polars::pl$DataFrame(
  destination = c("USA", "JPN", "BRA"),
  language = c("english", "japanese", "portuguese")
)

test

test2

cross_join(test, test2)
\dontshow{\}) # examplesIf}
}
