% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{describe}
\alias{describe}
\title{Summary statistics for a Polars DataFrame}
\usage{
describe(.data, percentiles = c(0.25, 0.75))
}
\arguments{
\item{percentiles}{One or more percentiles to include in the summary
statistics. All values must be between 0 and 1 (\code{NULL} are ignored).}
}
\description{
This function is deprecated as of tidypolars 0.10.0, it will be removed in
a future update. Use \code{summary()} with the same arguments instead.
}
