% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{describe_plan}
\alias{describe_plan}
\alias{describe_optimized_plan}
\title{Show the optimized and non-optimized query plans}
\usage{
describe_plan(.data)

describe_optimized_plan(.data)
}
\arguments{
\item{.data}{A Polars LazyFrame}
}
\description{
Those functions are deprecated as of tidypolars 0.10.0, they will be removed
in a future update. Use \code{explain()} with \code{optimized = FALSE} to recover the
output of \code{describe_plan()}, and with \code{optimized = TRUE} (the default) to
get the output of \code{describe_optimized_plan()}.
}
