% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distinct.R
\name{distinct.polars_data_frame}
\alias{distinct.polars_data_frame}
\alias{distinct.polars_lazy_frame}
\alias{duplicated_rows}
\title{Remove or keep only duplicated rows in a Data/LazyFrame}
\usage{
\method{distinct}{polars_data_frame}(.data, ..., keep = "first", maintain_order = TRUE)

\method{distinct}{polars_lazy_frame}(.data, ..., keep = "first", maintain_order = TRUE)

duplicated_rows(.data, ...)
}
\arguments{
\item{.data}{A Polars Data/LazyFrame}

\item{...}{Any expression accepted by \code{dplyr::select()}: variable names,
column numbers, select helpers, etc.}

\item{keep}{Either "first" (keep the first occurrence of the duplicated row),
"last" (last occurrence) or "none" (remove all ofccurences of duplicated
rows).}

\item{maintain_order}{Maintain row order. This is the default but it can
slow down the process with large datasets and it prevents the use of
streaming.}
}
\description{
By default, duplicates are looked for in all variables. It is possible to
specify a subset of variables where duplicates should be looked for. It is
also possible to keep either the first occurrence, the last occurence or
remove all duplicates.
}
\examples{
\dontshow{if (require("dplyr", quietly = TRUE) && require("tidyr", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
pl_test <- neopolars::pl$DataFrame(
  iso_o = c(rep(c("AA", "AB"), each = 2), "AC", "DC"),
  iso_d = rep(c("BA", "BB", "BC"), each = 2),
  value = c(2, 2, 3, 4, 5, 6)
)

distinct(pl_test)
distinct(pl_test, iso_o)

duplicated_rows(pl_test)
duplicated_rows(pl_test, iso_o, iso_d)
\dontshow{\}) # examplesIf}
}
