% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_na.R
\name{drop_na.DataFrame}
\alias{drop_na.DataFrame}
\alias{drop_na.LazyFrame}
\title{Drop missing values}
\usage{
\method{drop_na}{DataFrame}(data, ...)

\method{drop_na}{LazyFrame}(data, ...)
}
\arguments{
\item{data}{A Polars Data/LazyFrame}

\item{...}{Any expression accepted by \code{dplyr::select()}: variable names,
column numbers, select helpers, etc.}
}
\description{
By default, this will drop rows that contain any missing values. It is
possible to specify a subset of variables so that only missing values in these
variables will be considered.
}
\examples{
\dontshow{if (require("dplyr", quietly = TRUE) && require("tidyr", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
tmp <- mtcars
tmp[1:3, "mpg"] <- NA
tmp[4, "hp"] <- NA
pl_tmp <- polars::pl$DataFrame(tmp)

drop_na(pl_tmp)
drop_na(pl_tmp, hp, mpg)
\dontshow{\}) # examplesIf}
}
