% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain.R
\name{explain.RPolarsLazyFrame}
\alias{explain.RPolarsLazyFrame}
\title{Show the optimized and non-optimized query plans}
\usage{
\method{explain}{RPolarsLazyFrame}(x, optimized = TRUE, ...)
}
\arguments{
\item{x}{A Polars LazyFrame.}

\item{optimized}{Logical. If \code{TRUE} (default), show the query optimized by
Polars. Otherwise, show the initial query.}
}
\description{
This function is available for \code{LazyFrame}s only.

By default, \code{explain()} shows the query plan that is optimized and then run
by Polars. Setting \code{optimized = FALSE} shows the query plan as-is, without
any optimization done, but this not the query performed. Note that the
plans are read from bottom to top.
}
\examples{
\dontshow{if (require("dplyr", quietly = TRUE) && require("tidyr", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
query <- mtcars |>
  as_polars_lf() |>
  arrange(drat) |>
  filter(cyl == 3) |>
  select(mpg)

explain(query, optimized = FALSE)

explain(query)
\dontshow{\}) # examplesIf}
}
