% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter.DataFrame}
\alias{filter.DataFrame}
\alias{filter.LazyFrame}
\title{Keep rows that match a condition}
\usage{
\method{filter}{DataFrame}(.data, ..., .by = NULL)

\method{filter}{LazyFrame}(.data, ..., .by = NULL)
}
\arguments{
\item{.data}{A Polars Data/LazyFrame}

\item{...}{Expressions that return a logical value, and are defined in terms
of the variables in the data. If multiple expressions are included, they
will be combined with the & operator. Only rows for which all conditions
evaluate to \code{TRUE} are kept.}

\item{.by}{Optionally, a selection of columns to group by for just this
operation, functioning as an alternative to \code{group_by()}.}
}
\description{
This function is used to subset a data frame, retaining all rows that satisfy
your conditions. To be retained, the row must produce a value of TRUE for all
conditions. Note that when a condition evaluates to NA the row will be
dropped, unlike base subsetting with \code{[}.
}
\examples{
\dontshow{if (require("dplyr", quietly = TRUE) && require("tidyr", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
pl_iris <- polars::pl$DataFrame(iris)

filter(pl_iris, Sepal.Length < 5, Species == "setosa")

filter(pl_iris, Sepal.Length < Sepal.Width + Petal.Length)

filter(pl_iris, Species == "setosa" | is.na(Species))

filter(pl_iris, between(Sepal.Length, 5, 6, include_bounds = FALSE))

iris2 <- iris
iris2$Species <- as.character(iris2$Species)
iris2 |>
  as_polars() |>
  filter(Species \%in\% c("setosa", "virginica"))
\dontshow{\}) # examplesIf}
}
