% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-expr.R
\name{find_function_call_in_string}
\alias{find_function_call_in_string}
\title{In a deparsed expression, find the function calls and show which has an
equivalent in polars and which hasn't}
\usage{
find_function_call_in_string(x)
}
\description{
In a deparsed expression, find the function calls and show which has an
equivalent in polars and which hasn't
}
