% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_by.R
\name{group_by.DataFrame}
\alias{group_by.DataFrame}
\alias{ungroup.DataFrame}
\title{Group by one or more variables}
\usage{
\method{group_by}{DataFrame}(.data, ..., maintain_order = FALSE)

\method{ungroup}{DataFrame}(.data)
}
\arguments{
\item{.data}{A Polars Data/LazyFrame}

\item{...}{Variables to group by (used in \code{group_by()} only).}

\item{maintain_order}{Maintain row order. For performance reasons, this is
\code{FALSE} by default). Setting it to \code{TRUE} can slow down the process with
large datasets and prevents the use of streaming.}
}
\description{
Most data operations are done on groups defined by variables. \code{group_by()}
takes an existing Polars Data/LazyFrame and converts it into a grouped one
where operations are performed "by group". \code{ungroup()} removes grouping.
}
\examples{
by_cyl <- mtcars |>
  as_polars() |>
  group_by(cyl)

by_cyl

by_cyl |> summarise(
  disp = mean(disp),
  hp = mean(hp)
)
by_cyl |> filter(disp == max(disp))

}
