% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{group_by.RPolarsDataFrame}
\alias{group_by.RPolarsDataFrame}
\alias{ungroup.RPolarsDataFrame}
\alias{group_by.RPolarsLazyFrame}
\alias{ungroup.RPolarsLazyFrame}
\title{Group by one or more variables}
\usage{
\method{group_by}{RPolarsDataFrame}(.data, ..., maintain_order = FALSE)

\method{ungroup}{RPolarsDataFrame}(x, ...)

\method{group_by}{RPolarsLazyFrame}(.data, ..., maintain_order = FALSE)

\method{ungroup}{RPolarsLazyFrame}(x, ...)
}
\arguments{
\item{.data}{A Polars Data/LazyFrame}

\item{...}{Variables to group by (used in \code{group_by()} only). Not used in
\code{ungroup()}.}

\item{maintain_order}{Maintain row order. For performance reasons, this is
\code{FALSE} by default). Setting it to \code{TRUE} can slow down the process with
large datasets and prevents the use of streaming.}

\item{x}{A Polars Data/LazyFrame}
}
\description{
Most data operations are done on groups defined by variables. \code{group_by()}
takes an existing Polars Data/LazyFrame and converts it into a grouped one
where operations are performed "by group". \code{ungroup()} removes grouping.
}
\examples{
\dontshow{if (require("dplyr", quietly = TRUE) && require("tidyr", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
by_cyl <- mtcars |>
  as_polars_df() |>
  group_by(cyl)

by_cyl

by_cyl |> summarise(
  disp = mean(disp),
  hp = mean(hp)
)
by_cyl |> filter(disp == max(disp))
\dontshow{\}) # examplesIf}
}
