% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{group_split.polars_data_frame}
\alias{group_split.polars_data_frame}
\title{Grouping metadata}
\usage{
\method{group_split}{polars_data_frame}(.tbl, ..., .keep = TRUE)
}
\arguments{
\item{.tbl}{A Polars Data/LazyFrame}

\item{...}{If \code{.tbl} is not grouped, variables to group by. If \code{.tbl} is
already grouped, this is ignored.}

\item{.keep}{Should the grouping columns be kept?}
}
\description{
\code{group_vars()} returns a character vector with the names of the grouping
variables. \code{group_keys()} returns a data frame with one row per group.
}
\examples{
\dontshow{if (require("dplyr", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
pl_g <- polars::as_polars_df(iris) |>
  group_by(Species)

group_split(pl_g)
\dontshow{\}) # examplesIf}
}
