% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{group_vars.polars_data_frame}
\alias{group_vars.polars_data_frame}
\alias{group_vars.polars_lazy_frame}
\alias{group_keys.polars_data_frame}
\alias{group_keys.polars_lazy_frame}
\title{Grouping metadata}
\usage{
\method{group_vars}{polars_data_frame}(x)

\method{group_vars}{polars_lazy_frame}(x)

\method{group_keys}{polars_data_frame}(.tbl, ...)

\method{group_keys}{polars_lazy_frame}(.tbl, ...)
}
\arguments{
\item{x, .tbl}{A Polars Data/LazyFrame}

\item{...}{These dots are for future extensions and must be empty.}
}
\description{
\code{group_vars()} returns a character vector with the names of the grouping
variables. \code{group_keys()} returns a data frame with one row per group.
}
\examples{
\dontshow{if (require("dplyr", quietly = TRUE)) withAutoprint(\{ # examplesIf}
pl_g <- polars::as_polars_df(mtcars) |>
  group_by(cyl, am)

group_vars(pl_g)

group_keys(pl_g)
\dontshow{\}) # examplesIf}
}
