% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_vars.R
\name{group_vars.RPolarsDataFrame}
\alias{group_vars.RPolarsDataFrame}
\alias{group_vars.RPolarsLazyFrame}
\alias{group_keys.RPolarsDataFrame}
\alias{group_keys.RPolarsLazyFrame}
\title{Grouping metadata}
\usage{
\method{group_vars}{RPolarsDataFrame}(.data)

\method{group_vars}{RPolarsLazyFrame}(.data)

\method{group_keys}{RPolarsDataFrame}(.data)

\method{group_keys}{RPolarsLazyFrame}(.data)
}
\arguments{
\item{.data}{A Polars Data/LazyFrame}
}
\description{
\code{group_vars()} returns a character vector with the names of the grouping
variables. \code{group_keys()} returns a data frame with one row per group.
}
\examples{
pl_g <- polars::as_polars_df(mtcars) |>
  group_by(cyl, am)

group_vars(pl_g)

group_keys(pl_g)
}
