% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_unique_id.R
\name{make_unique_id}
\alias{make_unique_id}
\title{Create a column with unique id per row values}
\usage{
make_unique_id(.data, ..., new_col = "hash")
}
\arguments{
\item{.data}{A Polars Data/LazyFrame}

\item{...}{Any expression accepted by \code{dplyr::select()}: variable names,
column numbers, select helpers, etc.}

\item{new_col}{Name of the new column}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

The underlying Polars function isn't guaranteed to give the same results
across different versions. Therefore, this function will be removed and has
no replacement in \code{tidypolars}.
}
