% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.R
\name{pl_left_join}
\alias{pl_left_join}
\alias{pl_right_join}
\alias{pl_full_join}
\alias{pl_inner_join}
\title{Mutating joins}
\usage{
pl_left_join(x, y, by = NULL, suffix = c(".x", ".y"))

pl_right_join(x, y, by = NULL, suffix = c(".x", ".y"))

pl_full_join(x, y, by = NULL, suffix = c(".x", ".y"))

pl_inner_join(x, y, by = NULL, suffix = c(".x", ".y"))
}
\arguments{
\item{x, y}{Two Polars Data/LazyFrames}

\item{by}{Variables to join by. If \code{NULL}, the default, \verb{*_join()} will
perform a natural join, using all variables in common across \code{x} and \code{y}. A
message lists the variables so that you can check they're correct; suppress
the message by supplying \code{by} explicitly.}

\item{suffix}{If there are non-joined duplicate variables in \code{x} and \code{y},
these suffixes will be added to the output to disambiguate them. Should be a
character vector of length 2.}
}
\description{
Mutating joins add columns from \code{y} to \code{x}, matching observations based on
the keys.
}
\examples{
test <- polars::pl$DataFrame(
  x = c(1, 2, 3),
  y = c(1, 2, 3),
  z = c(1, 2, 3)
)

test2 <- polars::pl$DataFrame(
  x = c(1, 2, 4),
  y = c(1, 2, 4),
  z2 = c(4, 5, 7)
)

test

test2

pl_left_join(test, test2)

pl_inner_join(test, test2)

pl_full_join(test, test2)

#' Show how the arg 'suffix' works:
test2 <- polars::pl$DataFrame(
  x = c(1, 2, 4),
  y = c(1, 2, 4),
  z = c(4, 5, 7)
)

pl_left_join(test, test2, by = c("x", "y"), suffix = c("_left", "_right"))
}
