% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot.R
\name{pivot_longer.DataFrame}
\alias{pivot_longer.DataFrame}
\title{Pivot Data/LazyFrame from wide to long}
\usage{
\method{pivot_longer}{DataFrame}(data, cols, ..., names_to = "name", values_to = "value")
}
\arguments{
\item{data}{A Polars Data/LazyFrame}

\item{cols}{Columns to pivot into longer format. Can be anything accepted by
\code{dplyr::select()}.}

\item{...}{Not used.}

\item{names_to}{The (quoted) name of the column that will contain the column
names specified by \code{cols}.}

\item{values_to}{A string specifying the name of the column to create from the
data stored in cell values.}
}
\description{
Pivot Data/LazyFrame from wide to long
}
\examples{
\dontshow{if (require("dplyr", quietly = TRUE) && require("tidyr", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
pl_relig_income <- polars::pl$DataFrame(relig_income)
pl_relig_income

pl_relig_income |>
  pivot_longer(!religion, names_to = "income", values_to = "count")
\dontshow{\}) # examplesIf}
}
