% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_wider.R
\name{pivot_wider.polars_data_frame}
\alias{pivot_wider.polars_data_frame}
\title{Pivot a DataFrame from long to wide}
\usage{
\method{pivot_wider}{polars_data_frame}(
  data,
  ...,
  id_cols = NULL,
  names_from = name,
  values_from = value,
  names_prefix = "",
  names_sep = "_",
  names_glue = NULL,
  values_fill = NULL
)
}
\arguments{
\item{data}{A Polars DataFrame (LazyFrames are not supported).}

\item{...}{Dots which should be empty.}

\item{id_cols}{A set of columns that uniquely identify each observation.
Typically used when you have redundant variables, i.e. variables whose
values are perfectly correlated with existing variables.

Defaults to all columns in data except for the columns specified through
\code{names_from} and \code{values_from}. If a tidyselect expression is supplied, it
will be evaluated on data after removing the columns specified through
\code{names_from} and \code{values_from}.}

\item{names_from}{The (quoted or unquoted) column names whose values will be
used for the names of the new columns.}

\item{values_from}{The (quoted or unquoted) column names whose values will be
used to fill the new columns.}

\item{names_prefix}{String added to the start of every variable name. This is
particularly useful if \code{names_from} is a numeric vector and you want to
create syntactic variable names.}

\item{names_sep}{If \code{names_from} or \code{values_from} contains multiple
variables, this will be used to join their values together into a single
string to use as a column name.}

\item{names_glue}{Instead of \code{names_sep} and \code{names_prefix}, you can supply a
\code{glue} specification that uses the \code{names_from} columns to create custom
column names.}

\item{values_fill}{A scalar that will be used to replace missing values in
the new columns. Note that the type of this value will be applied to new
columns. For example, if you provide a character value to fill numeric
columns, then all these columns will be converted to character.}
}
\description{
Pivot a DataFrame from long to wide
}
\section{Unknown arguments}{


Arguments that are supported by the original implementation in the tidyverse
but are not listed above will throw a warning by default if they are
specified. To change this behavior to error instead, use
\code{options(tidypolars_unknown_args = "error")}.
}

\examples{
\dontshow{if (require("dplyr", quietly = TRUE) && require("tidyr", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
pl_fish_encounters <- as_polars_df(tidyr::fish_encounters)

pl_fish_encounters |>
  pivot_wider(names_from = station, values_from = seen)

pl_fish_encounters |>
  pivot_wider(names_from = station, values_from = seen, values_fill = 0)

# be careful about the type of the replacement value!
pl_fish_encounters |>
  pivot_wider(names_from = station, values_from = seen, values_fill = "a")

# using "names_glue" to specify the names of new columns
production <- expand.grid(
  product = c("A", "B"),
  country = c("AI", "EI"),
  year = 2000:2014
) |>
  filter((product == "A" & country == "AI") | product == "B") |>
  mutate(production = 1:45) |>
  as_polars_df()

production

production |>
  pivot_wider(
    names_from = c(product, country),
    values_from = production,
    names_glue = "prod_{product}_{country}"
  )
\dontshow{\}) # examplesIf}
}
