% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange.R
\name{pl_arrange}
\alias{pl_arrange}
\title{Order rows using column values}
\usage{
pl_arrange(.data, ..., .by_group = FALSE)
}
\arguments{
\item{.data}{A Polars Data/LazyFrame}

\item{...}{Quoted or unquoted variable names. Select helpers cannot be used.}

\item{.by_group}{If \code{TRUE}, will sort data within groups.}
}
\description{
Order rows using column values
}
\examples{
pl_test <- polars::pl$DataFrame(
  x1 = c("a", "a", "b", "a", "c"),
  x2 = c(2, 1, 5, 3, 1),
  value = sample(1:5)
)

pl_arrange(pl_test, x1)
pl_arrange(pl_test, x1, -x2)

if the data is grouped, you need to specify `.by_group = TRUE` to sort by
the groups first
pl_test |>
  pl_group_by(x1) |>
  pl_arrange(-x2, .by_group = TRUE)
}
