% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind.R
\name{pl_bind_cols}
\alias{pl_bind_cols}
\title{Append multiple Data/LazyFrames next to each other}
\usage{
pl_bind_cols(...)
}
\arguments{
\item{...}{Polars DataFrames or LazyFrames to combine. Each argument can
either be a Data/LazyFrame, or a list of Data/LazyFrames. Columns are matched
by name. All Data/LazyFrames must have the same number of rows and there
mustn't be duplicated column names.}
}
\description{
Append multiple Data/LazyFrames next to each other
}
\examples{
p1 <- polars::pl$DataFrame(
  x = sample(letters, 20),
  y = sample(1:100, 20)
)
p2 <- polars::pl$DataFrame(
  z = sample(letters, 20),
  w = sample(1:100, 20)
)

pl_bind_cols(p1, p2)
pl_bind_cols(list(p1, p2))
}
