% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind.R
\name{pl_bind_rows}
\alias{pl_bind_rows}
\title{Stack multiple Data/LazyFrames on top of each other}
\usage{
pl_bind_rows(...)
}
\arguments{
\item{...}{Polars DataFrames or LazyFrames to combine. Each argument can
either be a Data/LazyFrame, or a list of Data/LazyFrames. Columns are matched
by name. All Data/LazyFrames must have the same number of columns with
identical names.}
}
\description{
Stack multiple Data/LazyFrames on top of each other
}
\examples{
p1 <- polars::pl$DataFrame(
  x = sample(letters, 20),
  y = sample(1:100, 20)
)
p2 <- polars::pl$DataFrame(
  x = sample(letters, 20),
  y = sample(1:100, 20)
)

pl_bind_rows(p1, p2)

# this is equivalent
pl_bind_rows(list(p1, p2))
}
