% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-polars.R
\name{pl_colnames}
\alias{pl_colnames}
\title{Get column names of a Polars Data/LazyFrame}
\usage{
pl_colnames(x)
}
\arguments{
\item{x}{A Polars Data/LazyFrame}
}
\value{
A character vector with the column names
}
\description{
Get column names of a Polars Data/LazyFrame
}
