% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete.R
\name{pl_complete}
\alias{pl_complete}
\title{Complete a data frame with missing combinations of data}
\usage{
pl_complete(.data, ..., fill = list())
}
\arguments{
\item{.data}{A Polars Data/LazyFrame}

\item{...}{Any expression accepted by \code{dplyr::select()}: variable names,
column numbers, select helpers, etc.}

\item{fill}{A named list that for each variable supplies a single value to
use instead of \code{NA} for missing combinations.}
}
\description{
Turns implicit missing values into explicit missing values. This is useful
for completing missing combinations of data. Note that this function doesn't
work with grouped data yet.
}
\examples{
df <- polars::pl$DataFrame(
  group = c(1:2, 1, 2),
  item_id = c(1:2, 2, 3),
  item_name = c("a", "a", "b", "b"),
  value1 = c(1, NA, 3, 4),
  value2 = 4:7
)
df

df |> pl_complete(group, item_id, item_name)

df |>
  pl_complete(
    group, item_id, item_name,
    fill = list(value1 = 0, value2 = 99)
  )
}
