% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{pl_count}
\alias{pl_count}
\alias{pl_add_count}
\title{Count the observations in each group}
\usage{
pl_count(.data, ..., sort = FALSE, name = "n")

pl_add_count(.data, ..., sort = FALSE, name = "n")
}
\arguments{
\item{.data}{A Polars Data/LazyFrame}

\item{...}{Any expression accepted by \code{dplyr::select()}: variable names,
column numbers, select helpers, etc.}

\item{sort}{If \code{TRUE}, will show the largest groups at the top.}

\item{name}{Name of the new column.}
}
\description{
Count the observations in each group
}
\examples{
test <- polars::pl$DataFrame(mtcars)
pl_count(test, cyl)

pl_count(test, cyl, am)

pl_count(test, cyl, am, sort = TRUE, name = "count")

pl_add_count(test, cyl, am, sort = TRUE, name = "count")
}
