% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_na.R
\name{pl_drop_na}
\alias{pl_drop_na}
\title{Drop missing values}
\usage{
pl_drop_na(.data, ...)
}
\arguments{
\item{.data}{A Polars Data/LazyFrame}

\item{...}{Any expression accepted by \code{dplyr::select()}: variable names,
column numbers, select helpers, etc.}
}
\description{
By default, this will drop rows that contain any missing values. It is
possible to specify a subset of variables so that only missing values in these
variables will be considered.
}
\examples{
tmp <- mtcars
tmp[1:3, "mpg"] <- NA
tmp[4, "hp"] <- NA
pl_tmp <- polars::pl$DataFrame(tmp)

pl_drop_na(pl_tmp)
pl_drop_na(pl_tmp, hp, mpg)
}
