% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill.R
\name{pl_fill}
\alias{pl_fill}
\title{Fill in missing values with previous or next value}
\usage{
pl_fill(.data, ..., direction = c("down", "up", "downup", "updown"))
}
\arguments{
\item{.data}{A Polars Data/LazyFrame}

\item{direction}{Direction in which to fill missing values. Either "down"
(the default), "up", "downup" (i.e. first down and then up) or "updown"
(first up and then down).}
}
\description{
Fills missing values in selected columns using the next or previous entry.
This is useful in the common output format where values are not repeated, and
are only recorded when they change.
}
\details{
With grouped Data/LazyFrames, pl_fill() will be applied within each group,
meaning that it won't fill across group boundaries.
}
\examples{
pl_test <- polars::pl$DataFrame(x = c(NA, 1), y = c(2, NA))

pl_fill(pl_test, everything(), direction = "down")
pl_fill(pl_test, everything(), direction = "up")

# with grouped data, it doesn't use values from other groups
pl_grouped <- polars::pl$DataFrame(
  grp = rep(c("A", "B"), each = 3),
  x = c(1, NA, NA, NA, 2, NA),
  y = c(3, NA, 4, NA, 3, 1)
) |>
  group_by(grp)

pl_fill(pl_grouped, x, y, direction = "down")
}
