% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{pl_filter}
\alias{pl_filter}
\title{Keep rows that match a condition}
\usage{
pl_filter(data, ...)
}
\arguments{
\item{data}{A Polars Data/LazyFrame}

\item{...}{Expressions that return a logical value, and are defined in terms
of the variables in the data. If multiple expressions are included, they
will be combined with the & operator. Only rows for which all conditions
evaluate to \code{TRUE} are kept.}
}
\description{
This function is used to subset a data frame, retaining all rows that satisfy
your conditions. To be retained, the row must produce a value of TRUE for all
conditions. Note that when a condition evaluates to NA the row will be
dropped, unlike base subsetting with \code{[}.
}
\examples{
test <- polars::pl$DataFrame(iris)

pl_filter(pl_iris, Sepal.Length < 5 & Species == "setosa")

pl_filter(pl_iris, Sepal.Length < Sepal.Width + Petal.Length)

pl_filter(pl_iris, Species == "setosa" | is.na(Species))

iris2 <- iris
iris2$Species <- as.character(iris2$Species)
iris2 |>
  as_polars() |>
  pl_filter(Species \%in\% c("setosa", "virginica"))

pl_filter(pl_iris, between(Sepal.Length, 5, 6, include_bounds = FALSE))
}
