% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_by.R
\name{pl_group_by}
\alias{pl_group_by}
\alias{pl_ungroup}
\title{Group by one or more variables}
\usage{
pl_group_by(.data, ..., maintain_order = FALSE)

pl_ungroup(.data)
}
\arguments{
\item{.data}{A Polars Data/LazyFrame}

\item{...}{Variables to group by (used in \code{pl_group_by()} only).}

\item{maintain_order}{Maintain row order. For performance reasons, this is
\code{FALSE} by default). Setting it to \code{TRUE} can slow down the process with
large datasets and prevents the use of streaming.}
}
\description{
Most data operations are done on groups defined by variables. \code{pl_group_by()}
takes an existing Polars Data/LazyFrame and converts it into a grouped one
where operations are performed "by group". \code{pl_ungroup()} removes grouping.
}
\examples{
by_cyl <- mtcars |>
  as_polars() |>
  pl_group_by(cyl)

by_cyl

by_cyl |> pl_summarise(
  disp = mean(disp),
  hp = mean(hp)
)
by_cyl |> pl_filter(disp == max(disp))

}
