% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot.R
\name{pl_pivot_longer}
\alias{pl_pivot_longer}
\title{Pivot Data/LazyFrame from wide to long}
\usage{
pl_pivot_longer(data, cols, names_to = "name", values_to = "value")
}
\arguments{
\item{data}{A Polars Data/LazyFrame}

\item{cols}{Columns to pivot into longer format. Can be anything accepted by
`dplyr::select()`.}

\item{names_to}{The (quoted) name of the column that will contain the column
names specified by `cols`.}

\item{values_to}{A string specifying the name of the column to create from the
data stored in cell values.}
}
\description{
Pivot Data/LazyFrame from wide to long
}
