% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot.R
\name{pl_pivot_wider}
\alias{pl_pivot_wider}
\title{Pivot Data/LazyFrame from long to wide}
\usage{
pl_pivot_wider(.data, id_cols, names_from, values_from, values_fill = NULL)
}
\arguments{
\item{.data}{A Polars Data/LazyFrame}

\item{id_cols}{Defaults to all columns in data except for the columns specified
through \code{names_from} and \code{values_from}.}

\item{names_from}{The (quoted or unquoted) column name whose values will be
used for the names of the new columns.}

\item{values_from}{The (quoted or unquoted) column name whose values will be
used to fill the new columns.}

\item{values_fill}{A scalar that will be used to replace missing values in the
new columns. Note that the type of this value will be applied to new columns.
For example, if you provide a character value to fill numeric columns, then
all these columns will be converted to character.}
}
\description{
Pivot Data/LazyFrame from long to wide
}
\examples{
pl_fish_encounters <- polars::pl$DataFrame(fish_encounters)

pl_fish_encounters |>
  pl_pivot_wider(names_from = station, values_from = seen)

pl_fish_encounters |>
  pl_pivot_wider(names_from = station, values_from = seen, values_fill = 0)

# be careful about the type of the replacement value!
pl_fish_encounters |>
  pl_pivot_wider(names_from = station, values_from = seen, values_fill = "a")
}
