% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull.R
\name{pl_pull}
\alias{pl_pull}
\title{Extract a variable of a Data/LazyFrame}
\usage{
pl_pull(data, var)
}
\arguments{
\item{data}{A Polars Data/LazyFrame}

\item{var}{A quoted or unquoted variable name}
}
\description{
This returns an R vector and not a Polars Series.
}
\examples{
pl_test <- pl$DataFrame(iris)
pl_pull(pl_test, Sepal.Length)
pl_pull(pl_test, "Sepal.Length")
}
