% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relocate.R
\name{pl_relocate}
\alias{pl_relocate}
\title{Change column order}
\usage{
pl_relocate(.data, ..., .before = NULL, .after = NULL)
}
\arguments{
\item{.data}{A Polars Data/LazyFrame}

\item{...}{Any expression accepted by \code{dplyr::select()}: variable names,
column numbers, select helpers, etc.}

\item{.before, .after}{Column name (either quoted or unquoted) that
indicates the destination of columns selected by \code{...}. Supplying neither
will move columns to the left-hand side; specifying both is an error.}
}
\description{
Use \code{pl_relocate()} to change column positions, using the same syntax as
\code{pl_select()} to make it easy to move blocks of columns at once.
}
\examples{
dat <- as_polars(mtcars)

dat |>
  pl_relocate(hp, vs, .before = cyl)

# if .before and .after are not specified, selected columns are moved to the
# first positions
dat |>
  pl_relocate(hp, vs)

# .before and .after can be quoted or unquoted
dat |>
  pl_relocate(hp, vs, .after = "gear")

# select helpers are also available
dat |>
  pl_relocate(contains("[aeiou]"))
}
