% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_na.R
\name{pl_replace_na}
\alias{pl_replace_na}
\title{Replace NAs with specified values}
\usage{
pl_replace_na(data, replace)
}
\arguments{
\item{data}{A Polars Data/LazyFrame}

\item{replace}{Either a scalar that will be used to replace \code{NA} in all
columns, or a named list with the column name and the value that will be
used to replace \code{NA} in it. \strong{The column type will be automatically
converted to the type of the replacement value.}}
}
\description{
Replace NAs with specified values
}
\examples{
pl_test <- polars::pl$DataFrame(x = c(NA, 1), y = c(2, NA))

# replace all NA with 0
pl_replace_na(pl_test, 0)

# custom replacement per column
pl_replace_na(pl_test, list(x = 0, y = 999))

# be careful to use the same type for the replacement and for the column!
pl_replace_na(pl_test, list(x = "a", y = "unknown"))
}
