% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate.R
\name{pl_separate}
\alias{pl_separate}
\title{Separate a character column into multiple columns based on a substring}
\usage{
pl_separate(.data, col, into, sep = "[^[:alnum:]]+", remove = TRUE)
}
\arguments{
\item{.data}{A Polars Data/LazyFrame}

\item{col}{Column to split}

\item{into}{Character vector containing the names of new variables to create.
Use \code{NA} to omit the variable in the output.}

\item{sep}{String that is used to split the column.}

\item{remove}{If \code{TRUE}, remove input column from output data frame.}
}
\description{
Currently, splitting a column on a regular expression or position is not
possible.
}
\examples{
test <- polars::pl$DataFrame(
  x = c(NA, "x.y", "x.z", "y.z")
)
pl_separate(test, x, into = c("foo", "foo2"), sep = ".")
}
