% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice.R
\name{pl_slice_tail}
\alias{pl_slice_tail}
\alias{pl_slice_head}
\title{Subset the first or last rows of a Data/LazyFrame}
\usage{
pl_slice_tail(data, n = 5)

pl_slice_head(data, n = 5)
}
\arguments{
\item{data}{A Polars Data/LazyFrame}

\item{n}{The number of rows to select from the start or the end of the data.}
}
\description{
Subset the first or last rows of a Data/LazyFrame
}
\examples{
pl_test <- pl$DataFrame(iris)
pl_slice_head(pl_test, 3)
pl_slice_tail(pl_test, 3)
}
