% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unite.R
\name{pl_unite}
\alias{pl_unite}
\title{Unite multiple columns into one by pasting strings together}
\usage{
pl_unite(.data, col, ..., sep = "_", remove = TRUE, na.rm = FALSE)
}
\arguments{
\item{.data}{A Polars Data/LazyFrame}

\item{col}{The name of the new column, as a string or symbol.}

\item{sep}{Separator to use between values.}

\item{remove}{If \code{TRUE}, remove input columns from the output Data/LazyFrame.}

\item{na.rm}{If \code{TRUE}, missing values will be replaced with an empty string
prior to uniting each value.}
}
\description{
Unite multiple columns into one by pasting strings together
}
\examples{
test <- polars::pl$DataFrame(
  year = 2009:2011,
  month = 10:12,
  day = c(11L, 22L, 28L),
  name_day = c("Monday", "Thursday", "Wednesday")
)

# By default, united columns are dropped
pl_unite(test, col = "full_date", year, month, day, sep = "-")
pl_unite(test, col = "full_date", year, month, day, sep = "-", remove = FALSE)

test2 <- polars::pl$DataFrame(
  name = c("John", "Jack", "Thomas"),
  middlename = c("T.", NA, "F."),
  surname = c("Smith", "Thompson", "Jones")
)

# By default, NA values are kept in the character output
pl_unite(test2, col = "full_name", everything(), sep = " ")
pl_unite(test2, col = "full_name", everything(), sep = " ", na.rm = TRUE)
}
