% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.DataFrame}
\alias{print.DataFrame}
\title{Print method for DataFrame}
\usage{
\method{print}{DataFrame}(x, ...)
}
\arguments{
\item{x}{A Polars DataFrame}

\item{...}{Not used}
}
\description{
Print method for DataFrame
}
