% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull.R
\name{pull.DataFrame}
\alias{pull.DataFrame}
\title{Extract a variable of a Data/LazyFrame}
\usage{
\method{pull}{DataFrame}(.data, var, ...)
}
\arguments{
\item{.data}{A Polars Data/LazyFrame}

\item{var}{A quoted or unquoted variable name}

\item{...}{Not used.}
}
\description{
This returns an R vector and not a Polars Series.
}
\examples{
pl_test <- polars::pl$DataFrame(iris)
pull(pl_test, Sepal.Length)
pull(pl_test, "Sepal.Length")
}
