% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull.R
\name{pull.polars_data_frame}
\alias{pull.polars_data_frame}
\alias{pull.polars_lazy_frame}
\title{Extract a variable of a Data/LazyFrame}
\usage{
\method{pull}{polars_data_frame}(.data, var = -1, name = NULL, ...)

\method{pull}{polars_lazy_frame}(.data, var = -1, name = NULL, ...)
}
\arguments{
\item{.data}{A Polars Data/LazyFrame}

\item{var}{A quoted or unquoted variable name, or a variable index.}

\item{name}{An optional column to use to name the returned vector.}

\item{...}{Dots which should be empty.}
}
\description{
This returns an R vector and not a Polars Series.
}
\examples{
\dontshow{if (require("dplyr", quietly = TRUE)) withAutoprint(\{ # examplesIf}
pl_test <- as_polars_df(iris)
pull(pl_test, Sepal.Length)
pull(pl_test, "Sepal.Length")
\dontshow{\}) # examplesIf}
}
