% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{r_polars_funs}
\alias{r_polars_funs}
\title{List of functions in Polars and their equivalent in R}
\format{
An object of class \code{data.frame} with 195 rows and 3 columns.
}
\usage{
r_polars_funs
}
\description{
List of functions in Polars and their equivalent in R
}
\keyword{datasets}
