% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relocate.R
\name{relocate.DataFrame}
\alias{relocate.DataFrame}
\title{Change column order}
\usage{
\method{relocate}{DataFrame}(.data, ..., .before = NULL, .after = NULL)
}
\arguments{
\item{.data}{A Polars Data/LazyFrame}

\item{...}{Any expression accepted by \code{dplyr::select()}: variable names,
column numbers, select helpers, etc.}

\item{.before, .after}{Column name (either quoted or unquoted) that
indicates the destination of columns selected by \code{...}. Supplying neither
will move columns to the left-hand side; specifying both is an error.}
}
\description{
Use \code{relocate()} to change column positions, using the same syntax as
\code{select()} to make it easy to move blocks of columns at once.
}
\examples{
\dontshow{if (require("dplyr", quietly = TRUE) && require("tidyr", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
dat <- as_polars(mtcars)

dat |>
  relocate(hp, vs, .before = cyl)

# if .before and .after are not specified, selected columns are moved to the
# first positions
dat |>
  relocate(hp, vs)

# .before and .after can be quoted or unquoted
dat |>
  relocate(hp, vs, .after = "gear")

# select helpers are also available
dat |>
  relocate(contains("[aeiou]"))
\dontshow{\}) # examplesIf}
}
