% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.R
\name{rename.DataFrame}
\alias{rename.DataFrame}
\alias{rename_with.DataFrame}
\title{Rename columns}
\usage{
\method{rename}{DataFrame}(.data, ...)

\method{rename_with}{DataFrame}(.data, .fn, .cols = tidyselect::everything(), ...)
}
\arguments{
\item{.data}{A Polars Data/LazyFrame}

\item{...}{For \code{rename()}, one of the following:
\itemize{
\item params like \code{new_name = "old_name"} to rename selected variables.
\item as above but params wrapped in a list
}

For \code{rename_with}, additional arguments passed to \code{fn}.}

\item{.fn}{Function to apply on column names}

\item{.cols}{Columns on which to apply \code{fn}. Can be anything accepted by
\code{dplyr::select()}.}
}
\description{
Rename columns
}
\examples{
pl_test <- polars::pl$DataFrame(mtcars)

rename(pl_test, miles_per_gallon = "mpg", horsepower = "hp")

rename(pl_test, list(miles_per_gallon = "mpg", horsepower = "hp"))

rename(pl_test, `Miles per gallon` = "mpg", `Horse power` = "hp")

rename_with(pl_test, toupper, .cols = contains("p"))

pl_test_2 <- polars::pl$DataFrame(iris)

rename_with(pl_test_2, function(x) tolower(gsub(".", "_", x, fixed = TRUE)))

rename_with(pl_test_2, \(x) tolower(gsub(".", "_", x, fixed = TRUE)))
}
