% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_na.R
\name{replace_na.polars_data_frame}
\alias{replace_na.polars_data_frame}
\alias{replace_na.polars_lazy_frame}
\title{Replace NAs with specified values}
\usage{
\method{replace_na}{polars_data_frame}(data, replace, ...)

\method{replace_na}{polars_lazy_frame}(data, replace, ...)
}
\arguments{
\item{data}{A Polars Data/LazyFrame}

\item{replace}{Either a scalar that will be used to replace \code{NA} in all
columns, or a named list with the column name and the value that will be
used to replace \code{NA} in it.}

\item{...}{Dots which should be empty.}
}
\description{
Replace NAs with specified values
}
\examples{
\dontshow{if (require("dplyr", quietly = TRUE) && require("tidyr", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
pl_test <- neopolars::pl$DataFrame(x = c(NA, 1), y = c(2, NA))

# replace all NA with 0
replace_na(pl_test, 0)

# custom replacement per column
replace_na(pl_test, list(x = 0, y = 999))
\dontshow{\}) # examplesIf}
}
