% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-expr.R
\name{replace_vars_in_expr}
\alias{replace_vars_in_expr}
\title{In a deparsed expression, find the variable names and add pl$col() around
them}
\usage{
replace_vars_in_expr(data, deparsed)
}
\description{
In a deparsed expression, find the variable names and add pl$col() around
them
}
