% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select.DataFrame}
\alias{select.DataFrame}
\title{Select columns from a Data/LazyFrame}
\usage{
\method{select}{DataFrame}(.data, ...)
}
\arguments{
\item{.data}{A Polars Data/LazyFrame}

\item{...}{Any expression accepted by \code{dplyr::select()}: variable names,
column numbers, select helpers, etc.}
}
\description{
Select columns from a Data/LazyFrame
}
\examples{
\dontshow{if (require("dplyr", quietly = TRUE) && require("tidyr", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

pl_iris <- polars::pl$DataFrame(iris)

select(pl_iris, c("Sepal.Length", "Sepal.Width"))
select(pl_iris, Sepal.Length, Sepal.Width)
select(pl_iris, 1:3)
select(pl_iris, starts_with("Sepal"))
select(pl_iris, -ends_with("Length"))
\dontshow{\}) # examplesIf}
}
