% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate.R
\name{separate.polars_data_frame}
\alias{separate.polars_data_frame}
\alias{separate.polars_lazy_frame}
\title{Separate a character column into multiple columns based on a substring}
\usage{
\method{separate}{polars_data_frame}(data, col, into, sep = " ", remove = TRUE, ...)

\method{separate}{polars_lazy_frame}(data, col, into, sep = " ", remove = TRUE, ...)
}
\arguments{
\item{data}{A Polars Data/LazyFrame}

\item{col}{Column to split}

\item{into}{Character vector containing the names of new variables to create.
Use \code{NA} to omit the variable in the output.}

\item{sep}{String that is used to split the column. Regular expressions are
not supported yet.}

\item{remove}{If \code{TRUE}, remove input column from output data frame.}

\item{...}{Dots which should be empty.}
}
\description{
Currently, splitting a column on a regular expression or position is not
possible.
}
\examples{
\dontshow{if (require("dplyr", quietly = TRUE) && require("tidyr", quietly = TRUE)) withAutoprint(\{ # examplesIf}
test <- polars::pl$DataFrame(
  x = c(NA, "x.y", "x.z", "y.z")
)
separate(test, x, into = c("foo", "foo2"), sep = ".")
\dontshow{\}) # examplesIf}
}
