% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_query.R
\name{show_query.tidypolars}
\alias{show_query.tidypolars}
\title{Show the polars query equivalent to the tidyverse code}
\usage{
\method{show_query}{tidypolars}(x)
}
\arguments{
\item{x}{A Polars Data/LazyFrame}
}
\value{
Prints the polars code in the console and invisibly returns the
query as a character vector.
}
\description{
This function converts the tidyverse query into Polars code.
}
\examples{
\dontshow{if (require("dplyr", quietly = TRUE) && require("tidyr", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
tidy <- mtcars |>
  as_polars_df() |>
  arrange(drat, -cyl) |>
  add_count(am, gear) |>
  drop_na()

# the polars code is printed and assigned to `query`
query <- show_query(tidy)

query
\dontshow{\}) # examplesIf}
}
