% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice.R
\name{slice_tail.DataFrame}
\alias{slice_tail.DataFrame}
\alias{slice_head.DataFrame}
\alias{slice_sample.DataFrame}
\title{Subset rows of a Data/LazyFrame}
\usage{
\method{slice_tail}{DataFrame}(.data, ..., n)

\method{slice_head}{DataFrame}(.data, ..., n)

\method{slice_sample}{DataFrame}(.data, ..., n = NULL, prop = NULL, replace = FALSE)
}
\arguments{
\item{.data}{A Polars Data/LazyFrame}

\item{...}{Not used.}

\item{n}{The number of rows to select from the start or the end of the data.
Cannot be used with \code{prop}.}

\item{prop}{Proportion of rows to select. Cannot be used with \code{n}.}

\item{replace}{Perform the sampling with replacement (\code{TRUE}) or without
(\code{FALSE}).}
}
\description{
Subset rows of a Data/LazyFrame
}
\examples{
pl_test <- polars::pl$DataFrame(iris)
slice_head(pl_test, n = 3)
slice_tail(pl_test, n = 3)
slice_sample(pl_test, n = 5)
slice_sample(pl_test, prop = 0.1)
}
