% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice.R
\name{slice_tail.DataFrame}
\alias{slice_tail.DataFrame}
\alias{slice_tail.LazyFrame}
\alias{slice_head.DataFrame}
\alias{slice_head.LazyFrame}
\alias{slice_sample.DataFrame}
\title{Subset rows of a Data/LazyFrame}
\usage{
\method{slice_tail}{DataFrame}(.data, ..., n, by = NULL)

\method{slice_tail}{LazyFrame}(.data, ..., n, by = NULL)

\method{slice_head}{DataFrame}(.data, ..., n, by = NULL)

\method{slice_head}{LazyFrame}(.data, ..., n, by = NULL)

\method{slice_sample}{DataFrame}(.data, ..., n = NULL, prop = NULL, replace = FALSE, by = NULL)
}
\arguments{
\item{.data}{A Polars Data/LazyFrame}

\item{...}{Not used.}

\item{n}{The number of rows to select from the start or the end of the data.
Cannot be used with \code{prop}.}

\item{by}{Optionally, a selection of columns to group by for just this
operation, functioning as an alternative to \code{group_by()}. The group order
is not maintained, use \code{group_by()} if you want more control over it.}

\item{prop}{Proportion of rows to select. Cannot be used with \code{n}.}

\item{replace}{Perform the sampling with replacement (\code{TRUE}) or without
(\code{FALSE}).}
}
\description{
Subset rows of a Data/LazyFrame
}
\examples{
\dontshow{if (require("dplyr", quietly = TRUE) && require("tidyr", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
pl_test <- polars::pl$DataFrame(iris)
slice_head(pl_test, n = 3)
slice_tail(pl_test, n = 3)
slice_sample(pl_test, n = 5)
slice_sample(pl_test, prop = 0.1)
\dontshow{\}) # examplesIf}
}
