% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarize.polars_data_frame}
\alias{summarize.polars_data_frame}
\alias{summarise.polars_data_frame}
\alias{summarize.polars_lazy_frame}
\alias{summarise.polars_lazy_frame}
\title{Summarize each group down to one row}
\usage{
\method{summarize}{polars_data_frame}(.data, ..., .by = NULL, .groups = "drop_last")

\method{summarise}{polars_data_frame}(.data, ..., .by = NULL, .groups = "drop_last")

\method{summarize}{polars_lazy_frame}(.data, ..., .by = NULL, .groups = "drop_last")

\method{summarise}{polars_lazy_frame}(.data, ..., .by = NULL, .groups = "drop_last")
}
\arguments{
\item{.data}{A Polars Data/LazyFrame}

\item{...}{Name-value pairs. The name gives the name of the column in
the output. The value can be:
\itemize{
\item A vector the same length as the current group (or the whole data
frame if ungrouped).
\item NULL, to remove the column.
}

\code{across()} is mostly supported, except in a few cases. In particular, if the
\code{.cols} argument is \code{where(...)}, it will \emph{not} select variables that were
created before \code{across()}. Other select helpers are supported. See the examples.}

\item{.by}{Optionally, a selection of columns to group by for just this
operation, functioning as an alternative to \code{group_by()}. The group order
is not maintained, use \code{group_by()} if you want more control over it.}

\item{.groups}{Grouping structure of the result. Must be one of:
\itemize{
\item \code{"drop_last"} (default): drop the last level of grouping;
\item \code{"drop"}: all levels of grouping are dropped;
\item \code{"keep"}: keep the same grouping structure as \code{.data}.
}

For now, \code{"rowwise"} is not supported. Note that \code{dplyr} uses \code{.groups = NULL} by default, whose behavior depends on the number of rows by group in
the output. However, returning several rows by group in \code{summarize()} is
deprecated (one should use \code{reframe()} instead), which is why \code{.groups = NULL} is not supported by \code{tidypolars}.}
}
\description{
\code{summarize()} returns one row for each combination of grouping variables
(one difference with \code{dplyr::summarize()} is that \code{summarize()} only
accepts grouped data). It will contain one column for each grouping variable
and one column for each of the summary statistics that you have specified.
}
\examples{
\dontshow{if (require("dplyr", quietly = TRUE) && require("tidyr", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
mtcars |>
  as_polars_df() |>
  group_by(cyl) |>
  summarize(m_gear = mean(gear), sd_gear = sd(gear))

# an alternative syntax is to use `.by`
mtcars |>
  as_polars_df() |>
  summarize(m_gear = mean(gear), sd_gear = sd(gear), .by = cyl)
\dontshow{\}) # examplesIf}
}
