% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarize.DataFrame}
\alias{summarize.DataFrame}
\alias{summarise.DataFrame}
\title{Summarize each group down to one row}
\usage{
\method{summarize}{DataFrame}(.data, ...)

\method{summarise}{DataFrame}(.data, ...)
}
\arguments{
\item{.data}{A Polars Data/LazyFrame}
}
\description{
\code{summarize()} returns one row for each combination of grouping variables
(one difference with \code{dplyr::summarize()} is that \code{summarize()} only
accepts grouped data). It will contain one column for each grouping variable
and one column for each of the summary statistics that you have specified.
}
\examples{
mtcars |>
  as_polars() |>
  group_by(cyl) |>
  summarize(gear = mean(gear), gear2 = sd(gear))
}
