% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarize.RPolarsDataFrame}
\alias{summarize.RPolarsDataFrame}
\alias{summarise.RPolarsDataFrame}
\alias{summarize.RPolarsLazyFrame}
\alias{summarise.RPolarsLazyFrame}
\title{Summarize each group down to one row}
\usage{
\method{summarize}{RPolarsDataFrame}(.data, ..., .by = NULL)

\method{summarise}{RPolarsDataFrame}(.data, ..., .by = NULL)

\method{summarize}{RPolarsLazyFrame}(.data, ..., .by = NULL)

\method{summarise}{RPolarsLazyFrame}(.data, ..., .by = NULL)
}
\arguments{
\item{.data}{A Polars Data/LazyFrame}

\item{...}{Name-value pairs. The name gives the name of the column in
the output. The value can be:
\itemize{
\item A vector the same length as the current group (or the whole data
frame if ungrouped).
\item NULL, to remove the column.
}}

\item{.by}{Optionally, a selection of columns to group by for just this
operation, functioning as an alternative to \code{group_by()}. The group order
is not maintained, use \code{group_by()} if you want more control over it.}
}
\description{
\code{summarize()} returns one row for each combination of grouping variables
(one difference with \code{dplyr::summarize()} is that \code{summarize()} only
accepts grouped data). It will contain one column for each grouping variable
and one column for each of the summary statistics that you have specified.
}
\examples{
\dontshow{if (require("dplyr", quietly = TRUE) && require("tidyr", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
mtcars |>
  as_polars() |>
  group_by(cyl) |>
  summarize(m_gear = mean(gear), sd_gear = sd(gear))

# an alternative syntax is to use `.by`
mtcars |>
  as_polars() |>
  summarize(m_gear = mean(gear), sd_gear = sd(gear), .by = cyl)
\dontshow{\}) # examplesIf}
}
