% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.RPolarsDataFrame}
\alias{summary.RPolarsDataFrame}
\title{Summary statistics for a Polars DataFrame}
\usage{
\method{summary}{RPolarsDataFrame}(object, percentiles = c(0.25, 0.75), ...)
}
\arguments{
\item{object}{A Polars DataFrame.}

\item{percentiles}{One or more percentiles to include in the summary
statistics. All values must be between 0 and 1 (\code{NULL} are ignored).}

\item{...}{Ignored.}
}
\description{
Summary statistics for a Polars DataFrame
}
\examples{
mtcars |>
  as_polars_df() |>
  summary(percentiles = c(0.2, 0.4, 0.6, 0.8))
}
