% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidypolars-package.R
\docType{package}
\name{tidypolars-package}
\alias{tidypolars}
\alias{tidypolars-package}
\title{tidypolars: Get the Power of Polars with the Syntax of the Tidyverse}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Polars is a cross-language tool for manipulating very large data. However, one drawback is that the R implementation has a syntax that will look odd to many R users who are not used to Python syntax. The objective of tidypolars is to improve the ease-of-use of Polars in R by providing tidyverse syntax to polars.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://tidypolars.etiennebacher.com}
  \item \url{https://etiennebacher.r-universe.dev/tidypolars}
  \item Report bugs at \url{https://github.com/etiennebacher/tidypolars/issues}
}

}
\author{
\strong{Maintainer}: Etienne Bacher \email{etienne.bacher@protonmail.com} [copyright holder]

}
\keyword{internal}
