% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_r.R
\name{to_r}
\alias{to_r}
\title{Convert a Polars DataFrame to an R data.frame}
\usage{
to_r(.data)
}
\arguments{
\item{.data}{A Polars DataFrame only (LazyFrames are not modified). Any other
object is returned as-is.}
}
\description{
This is a simple wrapper of \verb{$to_data_frame()} present in \code{polars}. This is
only make this function work in a pipe workflow.
}
\examples{
iris |>
  as_polars() |>
  pl_filter(Sepal.Length > 6) |>
  to_r()
}
