% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncount.R
\name{uncount.DataFrame}
\alias{uncount.DataFrame}
\alias{uncount.LazyFrame}
\title{Uncount a Data/LazyFrame}
\usage{
\method{uncount}{DataFrame}(data, weights, ..., .remove = TRUE, .id = NULL)

\method{uncount}{LazyFrame}(data, weights, ..., .remove = TRUE, .id = NULL)
}
\arguments{
\item{data}{A Polars Data/LazyFrame}

\item{weights}{A vector of weights. Evaluated in the context of \code{data}.}

\item{...}{Not used.}

\item{.remove}{If \code{TRUE}, and weights is the name of a column in data, then
this column is removed.}

\item{.id}{Supply a string to create a new variable which gives a unique
identifier for each created row.}
}
\description{
This duplicates rows according to a weighting variable (or expression). This
is the opposite of \code{count()}.
}
\examples{
\dontshow{if (require("dplyr", quietly = TRUE) && require("tidyr", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
test <- pl$DataFrame(x = c("a", "b"), y = 100:101, n = c(1, 2))
test

uncount(test, n)

uncount(test, n, .id = "id")

# using constants
uncount(test, 2)

# using expressions
uncount(test, 2 / n)
\dontshow{\}) # examplesIf}
}
