% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unite.R
\name{unite.DataFrame}
\alias{unite.DataFrame}
\title{Unite multiple columns into one by pasting strings together}
\usage{
\method{unite}{DataFrame}(data, col, ..., sep = "_", remove = TRUE, na.rm = FALSE)
}
\arguments{
\item{data}{A Polars Data/LazyFrame}

\item{col}{The name of the new column, as a string or symbol.}

\item{...}{Any expression accepted by \code{dplyr::select()}: variable names,
column numbers, select helpers, etc.}

\item{sep}{Separator to use between values.}

\item{remove}{If \code{TRUE}, remove input columns from the output Data/LazyFrame.}

\item{na.rm}{If \code{TRUE}, missing values will be replaced with an empty string
prior to uniting each value.}
}
\description{
Unite multiple columns into one by pasting strings together
}
\examples{
\dontshow{if (require("dplyr", quietly = TRUE) && require("tidyr", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
test <- polars::pl$DataFrame(
  year = 2009:2011,
  month = 10:12,
  day = c(11L, 22L, 28L),
  name_day = c("Monday", "Thursday", "Wednesday")
)

# By default, united columns are dropped
unite(test, col = "full_date", year, month, day, sep = "-")
unite(test, col = "full_date", year, month, day, sep = "-", remove = FALSE)

test2 <- polars::pl$DataFrame(
  name = c("John", "Jack", "Thomas"),
  middlename = c("T.", NA, "F."),
  surname = c("Smith", "Thompson", "Jones")
)

# By default, NA values are kept in the character output
unite(test2, col = "full_name", everything(), sep = " ")
unite(test2, col = "full_name", everything(), sep = " ", na.rm = TRUE)
\dontshow{\}) # examplesIf}
}
