% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_csv_polars}
\alias{write_csv_polars}
\title{Export data to CSV file(s)}
\usage{
write_csv_polars(
  .data,
  file,
  ...,
  include_bom = FALSE,
  include_header = TRUE,
  separator = ",",
  line_terminator = "\\n",
  quote_char = "\\"",
  batch_size = 1024,
  datetime_format = NULL,
  date_format = NULL,
  time_format = NULL,
  float_precision = NULL,
  null_value = "",
  quote_style = "necessary"
)
}
\arguments{
\item{.data}{A Polars DataFrame.}

\item{file}{File path to which the result should be written.}

\item{...}{Ignored.}

\item{include_bom}{Whether to include UTF-8 BOM (byte order mark) in the CSV
output.}

\item{include_header}{Whether to include header in the CSV output.}

\item{separator}{Separate CSV fields with this symbol.}

\item{line_terminator}{String used to end each row.}

\item{quote_char}{Byte to use as quoting character.}

\item{batch_size}{Number of rows that will be processed per thread.}

\item{datetime_format}{A format string, with the specifiers defined by the
chrono Rust crate. If no format specified, the default fractional-second
precision is inferred from the maximum timeunit found in the frame’s Datetime
cols (if any).}

\item{date_format}{A format string, with the specifiers defined by the chrono
Rust crate.}

\item{time_format}{A format string, with the specifiers defined by the chrono
Rust crate.}

\item{float_precision}{Number of decimal places to write, applied to both
Float32 and Float64 datatypes.}

\item{null_value}{A string representing null values (defaulting to the empty
string).}

\item{quote_style}{Determines the quoting strategy used.
\itemize{
\item \code{"necessary"} (default): This puts quotes around fields only when necessary.
They are necessary when fields contain a quote, delimiter or record
terminator. Quotes are also necessary when writing an empty record (which
is indistinguishable from a record with one empty field). This is the
default.
\item \code{"always"}: This puts quotes around every field.
\item \code{"non_numeric"}: This puts quotes around all fields that are non-numeric.
Namely, when writing a field that does not parse as a valid float or integer,
then quotes will be used even if they aren`t strictly necessary.
\item \code{"never"}: This never puts quotes around fields, even if that results in
invalid CSV data (e.g. by not quoting strings containing the separator).
}}
}
\value{
The input DataFrame.
}
\description{
Export data to CSV file(s)
}
\details{
\subsection{Partitioned output}{

It is possible to export data to multiple files based on various parameters,
such as the values of some variables, or such that each file has a maximum
number of rows. See \code{\link[=partition_by]{partition_by()}} for more details.
}
}
\examples{
dest <- tempfile(fileext = ".csv")
mtcars |>
  as_polars_df() |>
  write_csv_polars(dest)

read.csv(dest)
}
