% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_json_polars}
\alias{write_json_polars}
\title{Export data to JSON file(s)}
\usage{
write_json_polars(.data, file)
}
\arguments{
\item{.data}{A Polars DataFrame.}

\item{file}{File path to which the result should be written.}
}
\value{
The input DataFrame.
}
\description{
Export data to JSON file(s)
}
\examples{
\dontshow{if (requireNamespace("jsonlite")) withAutoprint(\{ # examplesIf}
dest <- tempfile(fileext = ".json")
mtcars |>
  as_polars_df() |>
  write_json_polars(dest)

jsonlite::fromJSON(dest)
\dontshow{\}) # examplesIf}
}
