% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_ndjson_polars}
\alias{write_ndjson_polars}
\title{Export data to NDJSON file(s)}
\usage{
write_ndjson_polars(.data, file)
}
\arguments{
\item{.data}{A Polars DataFrame.}

\item{file}{File path to which the result should be written.}
}
\value{
The input DataFrame.
}
\description{
Export data to NDJSON file(s)
}
\details{
\subsection{Partitioned output}{

It is possible to export data to multiple files based on various parameters,
such as the values of some variables, or such that each file has a maximum
number of rows. See \code{\link[=partition_by]{partition_by()}} for more details.
}
}
\examples{
\dontshow{if (requireNamespace("jsonlite")) withAutoprint(\{ # examplesIf}
dest <- tempfile(fileext = ".ndjson")
mtcars |>
  as_polars_df() |>
  write_ndjson_polars(dest)

jsonlite::stream_in(file(dest), verbose = FALSE)
\dontshow{\}) # examplesIf}
}
