% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_parquet_polars}
\alias{write_parquet_polars}
\title{Export data to parquet file(s)}
\usage{
write_parquet_polars(
  source,
  ...,
  n_rows = NULL,
  row_index_name = NULL,
  row_index_offset = 0L,
  parallel = "auto",
  hive_partitioning = TRUE,
  glob = TRUE,
  rechunk = TRUE,
  low_memory = FALSE,
  storage_options = NULL,
  use_statistics = TRUE,
  cache = TRUE
)
}
\description{
Export data to parquet file(s)
}
