% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{describe}
\alias{describe}
\title{Summary statistics for a Polars DataFrame}
\usage{
describe(.data, percentiles = c(0.25, 0.75))
}
\arguments{
\item{.data}{A Polars Data/LazyFrame}

\item{percentiles}{One or more percentiles to include in the summary
statistics. All values must be between 0 and 1 (\code{NULL} are ignored).}
}
\description{
Summary statistics for a Polars DataFrame
}
\examples{
\dontshow{if (require("dplyr", quietly = TRUE) && require("tidyr", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
mtcars |>
  as_polars_df() |>
  describe(percentiles = c(0.2, 0.4, 0.6, 0.8))
\dontshow{\}) # examplesIf}
}
