% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{describe_plan}
\alias{describe_plan}
\alias{describe_optimized_plan}
\title{Show the optimized and non-optimized query plans}
\usage{
describe_plan(.data)

describe_optimized_plan(.data)
}
\arguments{
\item{.data}{A Polars LazyFrame}
}
\description{
These functions are available for \code{LazyFrame}s only. \code{describe_plan()} shows
the query plan as-is, without any optimization done by Polars. This is not
the query performed. Before running the query, Polars applies a list of
optimizations (such as filtering data before sorting it). The actual query
plan ran by Polars is shown by \code{describe_optimized_plan()}. Note that the
queries are read from bottom to top.
}
\examples{
query <- mtcars |>
  as_polars(lazy = TRUE) |>
  arrange(drat) |>
  filter(cyl == 3) |>
  select(mpg)

describe_plan(query)

describe_optimized_plan(query)
}
