% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch}
\alias{fetch}
\title{Fetch \code{n} rows of a LazyFrame}
\usage{
fetch(
  .data,
  n_rows = 500,
  type_coercion = TRUE,
  predicate_pushdown = TRUE,
  projection_pushdown = TRUE,
  simplify_expression = TRUE,
  slice_pushdown = TRUE,
  comm_subplan_elim = TRUE,
  comm_subexpr_elim = TRUE,
  cluster_with_columns = TRUE,
  no_optimization = FALSE,
  engine = c("auto", "in-memory", "streaming"),
  streaming = FALSE
)
}
\arguments{
\item{.data}{A Polars LazyFrame}

\item{n_rows}{Number of rows to fetch.}

\item{type_coercion}{Coerce types such that operations succeed and run on
minimal required memory (default is \code{TRUE}).}

\item{predicate_pushdown}{Applies filters as early as possible at scan level
(default is \code{TRUE}).}

\item{projection_pushdown}{Select only the columns that are needed at the scan
level (default is \code{TRUE}).}

\item{simplify_expression}{Various optimizations, such as constant folding
and replacing expensive operations with faster alternatives (default is
\code{TRUE}).}

\item{slice_pushdown}{Only load the required slice from the scan. Don't
materialize sliced outputs level. Don't materialize sliced outputs (default
is \code{TRUE}).}

\item{comm_subplan_elim}{Cache branching subplans that occur on self-joins or
unions (default is \code{TRUE}).}

\item{comm_subexpr_elim}{Cache common subexpressions (default is \code{TRUE}).}

\item{cluster_with_columns}{Combine sequential independent calls to
\verb{$with_columns()}.}

\item{no_optimization}{Sets the following optimizations to \code{FALSE}:
\code{predicate_pushdown}, \code{projection_pushdown},  \code{slice_pushdown},
\code{simplify_expression}. Default is \code{FALSE}.}

\item{engine}{The engine name to use for processing the query. One of the
followings:
\itemize{
\item \code{"auto"} (default): Select the engine automatically. The \code{"in-memory"}
engine will be selected for most cases.
\item \code{"in-memory"}: Use the in-memory engine.
\item \code{"streaming"}: \link{Experimental} Use the (new) streaming engine.
}}

\item{streaming}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Deprecated, use \code{engine}
instead.}
}
\description{
Fetch is a way to collect only the first \code{n} rows of a LazyFrame. It is
mainly used to test that a query runs as expected on a subset of the data
before using \code{collect()} on the full query. Note that fetching \code{n} rows
doesn't mean that the output will actually contain \code{n} rows, see the section
'Details' for more information.
}
\details{
The parameter \code{n_rows} indicates how many rows from the LazyFrame should be
used at the beginning of the query, but it doesn't guarantee that \code{n_rows} will
be returned. For example, if the query contains a filter or join operations
with other datasets, then the final number of rows can be lower than \code{n_rows}.
On the other hand, appending some rows during the query can lead to an output
that has more rows than \code{n_rows}.
}
\examples{
\dontshow{if (require("dplyr", quietly = TRUE) && require("tidyr", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
dat_lazy <- polars::as_polars_df(iris)$lazy()

# this will return 30 rows
fetch(dat_lazy, 30)

# this will return less than 30 rows because there are less than 30 matches
# for this filter in the whole dataset
dat_lazy |>
  filter(Sepal.Length > 7.0) |>
  fetch(30)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=collect]{collect()}} for applying a lazy query on the full data.
}
