% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.R
\name{pl_semi_join}
\alias{pl_semi_join}
\alias{pl_anti_join}
\title{Filtering joins}
\usage{
pl_semi_join(x, y, by = NULL)

pl_anti_join(x, y, by = NULL)
}
\arguments{
\item{x, y}{Two Polars Data/LazyFrames}

\item{by}{Variables to join by. If \code{NULL}, the default, \verb{*_join()} will
perform a natural join, using all variables in common across \code{x} and \code{y}. A
message lists the variables so that you can check they're correct; suppress
the message by supplying \code{by} explicitly.}
}
\description{
Filtering joins filter rows from \code{x} based on the presence or absence of
matches in \code{y}:
\itemize{
\item \code{pl_semi_join()} return all rows from \code{x} with a match in \code{y}.
\item \code{pl_anti_join()} return all rows from \code{x} without a match in \code{y}.
}
}
\examples{
test <- polars::pl$DataFrame(
  x = c(1, 2, 3),
  y = c(1, 2, 3),
  z = c(1, 2, 3)
)
test2 <- polars::pl$DataFrame(
  x = c(1, 2, 4),
  y = c(1, 2, 4),
  z2 = c(1, 2, 4)
)
test
test2
# only keep the rows of `test` that have matching keys in `test2`
pl_semi_join(test, test2, by = c("x", "y"))
# only keep the rows of `test` that don't have matching keys in `test2`
pl_anti_join(test, test2, by = c("x", "y"))
}
