% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fish_encounters}
\alias{fish_encounters}
\title{Fish encounters}
\format{
A dataset with variables:
\itemize{
\item fish: Fish identifier
\item station: Measurement station
\item seen: Was the fish seen? (1 if yes, and true for all rows)
}
}
\source{
Dataset provided by Myfanwy Johnston; more details at https://fishsciences.github.io/post/visualizing-fish-encounter-histories/
}
\usage{
fish_encounters
}
\description{
This dataset is reexported from \code{tidyr::fish_encounters}
}
\keyword{datasets}
