% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_scan.R
\name{from_ipc}
\alias{from_ipc}
\alias{read_ipc_polars}
\alias{scan_ipc_polars}
\title{Import data from IPC file(s)}
\usage{
read_ipc_polars(
  source,
  ...,
  n_rows = NULL,
  memory_map = TRUE,
  row_index_name = NULL,
  row_index_offset = 0L,
  rechunk = FALSE,
  cache = TRUE,
  include_file_paths = NULL
)

scan_ipc_polars(
  source,
  ...,
  n_rows = NULL,
  memory_map = TRUE,
  row_index_name = NULL,
  row_index_offset = 0L,
  rechunk = FALSE,
  cache = TRUE,
  include_file_paths = NULL
)
}
\arguments{
\item{source}{Path to a file. You can use globbing with \code{*} to scan/read multiple
files in the same directory (see examples).}

\item{...}{Ignored.}

\item{n_rows}{Maximum number of rows to read.}

\item{memory_map}{A logical. If \code{TRUE}, try to memory map the file.
This can greatly improve performance on repeated queries as the OS may cache pages.
Only uncompressed Arrow IPC files can be memory mapped.}

\item{row_index_name}{If not \code{NULL}, this will insert a row index column with
the given name into the DataFrame.}

\item{row_index_offset}{Offset to start the row index column (only used if
the name is set).}

\item{rechunk}{In case of reading multiple files via a glob pattern, rechunk
the final DataFrame into contiguous memory chunks.}

\item{cache}{Cache the result after reading.}

\item{include_file_paths}{Character value indicating the column name that will
include the path of the source file(s).}
}
\description{
\code{read_ipc_polars()} imports the data as a Polars DataFrame.

\code{scan_ipc_polars()} imports the data as a Polars LazyFrame.
}
\details{
Hive-style partitioning is not supported yet.
}
