% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{group_vars.RPolarsDataFrame}
\alias{group_vars.RPolarsDataFrame}
\alias{group_vars.RPolarsLazyFrame}
\alias{group_keys.RPolarsDataFrame}
\alias{group_keys.RPolarsLazyFrame}
\title{Grouping metadata}
\usage{
\method{group_vars}{RPolarsDataFrame}(x)

\method{group_vars}{RPolarsLazyFrame}(x)

\method{group_keys}{RPolarsDataFrame}(.tbl)

\method{group_keys}{RPolarsLazyFrame}(.tbl)
}
\arguments{
\item{x, .tbl}{A Polars Data/LazyFrame}
}
\description{
\code{group_vars()} returns a character vector with the names of the grouping
variables. \code{group_keys()} returns a data frame with one row per group.
}
\examples{
\dontshow{if (require("dplyr", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
pl_g <- polars::as_polars_df(mtcars) |>
  group_by(cyl, am)

group_vars(pl_g)

group_keys(pl_g)
\dontshow{\}) # examplesIf}
}
