% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.R
\name{left_join.DataFrame}
\alias{left_join.DataFrame}
\alias{right_join.DataFrame}
\alias{full_join.DataFrame}
\alias{inner_join.DataFrame}
\title{Mutating joins}
\usage{
\method{left_join}{DataFrame}(
  x,
  y,
  by = NULL,
  copy = NULL,
  suffix = c(".x", ".y"),
  ...,
  keep = NULL
)

\method{right_join}{DataFrame}(
  x,
  y,
  by = NULL,
  copy = NULL,
  suffix = c(".x", ".y"),
  ...,
  keep = NULL
)

\method{full_join}{DataFrame}(
  x,
  y,
  by = NULL,
  copy = NULL,
  suffix = c(".x", ".y"),
  ...,
  keep = NULL
)

\method{inner_join}{DataFrame}(
  x,
  y,
  by = NULL,
  copy = NULL,
  suffix = c(".x", ".y"),
  ...,
  keep = NULL
)
}
\arguments{
\item{x, y}{Two Polars Data/LazyFrames}

\item{by}{Variables to join by. If \code{NULL}, the default, \verb{*_join()} will
perform a natural join, using all variables in common across \code{x} and \code{y}. A
message lists the variables so that you can check they're correct; suppress
the message by supplying \code{by} explicitly.}

\item{copy, keep}{Not used.}

\item{suffix}{If there are non-joined duplicate variables in \code{x} and \code{y},
these suffixes will be added to the output to disambiguate them. Should be a
character vector of length 2.}

\item{...}{Not used.}
}
\description{
Mutating joins add columns from \code{y} to \code{x}, matching observations based on
the keys.
}
\examples{
test <- polars::pl$DataFrame(
  x = c(1, 2, 3),
  y = c(1, 2, 3),
  z = c(1, 2, 3)
)

test2 <- polars::pl$DataFrame(
  x = c(1, 2, 4),
  y = c(1, 2, 4),
  z2 = c(4, 5, 7)
)

test

test2

left_join(test, test2)

inner_join(test, test2)

full_join(test, test2)

# Show how the arg 'suffix' works:
test2 <- polars::pl$DataFrame(
  x = c(1, 2, 4),
  y = c(1, 2, 4),
  z = c(4, 5, 7)
)

left_join(test, test2, by = c("x", "y"), suffix = c("_left", "_right"))
}
