% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_unique_id.R
\name{make_unique_id}
\alias{make_unique_id}
\title{Create a column with unique id per row values}
\usage{
make_unique_id(.data, ..., new_col = "hash")
}
\arguments{
\item{.data}{A Polars Data/LazyFrame}

\item{...}{Any expression accepted by \code{dplyr::select()}: variable names,
column numbers, select helpers, etc.}

\item{new_col}{Name of the new column}
}
\description{
Create a column with unique id per row values
}
\examples{
\dontshow{if (require("dplyr", quietly = TRUE) && require("tidyr", quietly = TRUE)) withAutoprint(\{ # examplesIf}
mtcars |>
  as_polars_df() |>
  make_unique_id(am, gear)
\dontshow{\}) # examplesIf}
}
