% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sink.R
\name{partitioned_output}
\alias{partitioned_output}
\alias{partition_by_key}
\alias{partition_by_max_size}
\title{Helper functions to export a LazyFrame as a partitioned output}
\usage{
partition_by_key(
  base_path,
  ...,
  by,
  include_key = TRUE,
  per_partition_sort_by = NULL
)

partition_by_max_size(base_path, ..., max_size, per_partition_sort_by = NULL)
}
\arguments{
\item{base_path}{The base path for the output files. Use the \code{mkdir} option
of the \verb{sink_*} methods to ensure directories in the path are created.}

\item{...}{These dots are for future extensions and must be empty.}

\item{by}{Something can be coerced to a list of Polars expressions. Used to
partition by.}

\item{include_key}{If \code{TRUE} (default), include the key columns in the output
files.}

\item{per_partition_sort_by}{Something can be coerced to a list of Polars
expressions, or \code{NULL} (default). Used  to sort over within each partition.
Note that this might increase the memory consumption needed for each partition.}

\item{max_size}{An integer-ish value indicating the maximum number of rows in
each of the generated files.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
More details and examples in the documentation of \verb{sink_*()} functions.
}
