% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_longer.R
\name{pivot_longer.polars_data_frame}
\alias{pivot_longer.polars_data_frame}
\alias{pivot_longer.polars_lazy_frame}
\title{Pivot a Data/LazyFrame from wide to long}
\usage{
\method{pivot_longer}{polars_data_frame}(
  data,
  cols,
  ...,
  names_to = "name",
  names_prefix = NULL,
  values_to = "value"
)

\method{pivot_longer}{polars_lazy_frame}(
  data,
  cols,
  ...,
  names_to = "name",
  names_prefix = NULL,
  values_to = "value"
)
}
\arguments{
\item{data}{A Polars Data/LazyFrame}

\item{cols}{Columns to pivot into longer format. Can be anything accepted by
\code{dplyr::select()}.}

\item{...}{Dots which should be empty.}

\item{names_to}{The (quoted) name of the column that will contain the column
names specified by \code{cols}.}

\item{names_prefix}{A regular expression used to remove matching text from
the start of each variable name.}

\item{values_to}{A string specifying the name of the column to create from the
data stored in cell values.}
}
\description{
Pivot a Data/LazyFrame from wide to long
}
\section{Unknown arguments}{


Arguments that are supported by the original implementation in the tidyverse
but are not listed above will throw a warning by default if they are
specified. To change this behavior to error instead, use
\code{options(tidypolars_unknown_args = "error")}.
}

\examples{
\dontshow{if (require("dplyr", quietly = TRUE) && require("tidyr", quietly = TRUE)) withAutoprint(\{ # examplesIf}
pl_relig_income <- as_polars_df(tidyr::relig_income)
pl_relig_income

pl_relig_income |>
  pivot_longer(!religion, names_to = "income", values_to = "count")

pl_billboard <- as_polars_df(tidyr::billboard)
pl_billboard

pl_billboard |>
  pivot_longer(
    cols = starts_with("wk"),
    names_to = "week",
    names_prefix = "wk",
    values_to = "rank",
  )
\dontshow{\}) # examplesIf}
}
