% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange.R
\name{pl_arrange}
\alias{pl_arrange}
\title{Order rows using column values}
\usage{
pl_arrange(data, ...)
}
\arguments{
\item{data}{A Polars Data/LazyFrame}

\item{...}{Quoted or unquoted variable names. Select helpers cannot be used.}
}
\description{
Order rows using column values
}
\examples{
pl_test <- polars::pl$DataFrame(
  x1 = c("a", "a", "b", "a", "c"),
  x2 = c(2, 1, 5, 3, 1),
  value = sample(1:5)
)

pl_arrange(test, x1)
pl_arrange(test, -"x1")
pl_arrange(test, x1, -x2)
}
