% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{pl_collect}
\alias{pl_collect}
\title{Collect a LazyFrame}
\usage{
pl_collect(data)
}
\arguments{
\item{data}{A Polars LazyFrame}
}
\description{
Polars LazyFrames are not loaded in memory. Running `collect()` checks the
execution plan, optimizes it in the background and performs it. The result
is loaded in the R session.
}
\examples{
dat_lazy <- pl$DataFrame(iris)$lazy()
pl_collect(dat_lazy)
}
