% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete.R
\name{pl_complete}
\alias{pl_complete}
\title{Complete a data frame with missing combinations of data}
\usage{
pl_complete(data, ...)
}
\arguments{
\item{data}{A Polars Data/LazyFrame}

\item{...}{Any expression accepted by \code{dplyr::select()}: variable names,
column numbers, select helpers, etc.}
}
\description{
Turns implicit missing values into explicit missing values. This is useful
for completing missing combinations of data. Note that this function doesn't
work with grouped data yet.
}
\examples{
\dontshow{if (packageVersion("polars") >= "0.6.2") (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
test <- polars::pl$DataFrame(
  country = c("France", "France", "UK", "UK", "Spain"),
  year = c(2020, 2021, 2019, 2020, 2022),
  value = c(1, 2, 3, 4, 5)
)
test

pl_complete(test, country, year)
\dontshow{\}) # examplesIf}
}
