% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distinct.R
\name{pl_distinct}
\alias{pl_distinct}
\title{Remove duplicated rows in a Data/LazyFrame}
\usage{
pl_distinct(data, ..., keep = "first", maintain_order = TRUE)
}
\arguments{
\item{data}{A Polars Data/LazyFrame}

\item{...}{Any expression accepted by `dplyr::select()`: variable names,
column numbers, select helpers, etc.}

\item{keep}{Either "first" (keep the first occurrence of the duplicated row),
"last" (last occurrence) or "none" (remove all ofccurences of duplicated
rows).}

\item{maintain_order}{Maintain row order. This is the default but it can
slow down the process with large datasets and prevents the use of streaming.}
}
\description{
By default, duplicates are looked for in all variables. It is possible to
specify a subset of variables where duplicates should be looked for. It is
also possible to keep either the first occurrence, the last occurence or
remove all duplicates.
}
\examples{
pl_test <- pl$DataFrame(
  iso_o = rep(c("AA", "AB", "AC"), each = 2),
  iso_d = rep(c("BA", "BB", "BC"), each = 2),
  value = 1:6
)

pl_distinct(pl_test)
pl_distinct(pl_test, iso_o)
}
