% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{pl_summarize}
\alias{pl_summarize}
\alias{pl_summarise}
\title{Summarize each group down to one row}
\usage{
pl_summarize(.data, ...)

pl_summarise(.data, ...)
}
\arguments{
\item{.data}{A Polars Data/LazyFrame}

\item{...}{Name-value pairs. The name gives the name of the column in
the output. The value can be:
\itemize{
\item A vector the same length as the current group (or the whole data
frame if ungrouped).
\item NULL, to remove the column.
}}
}
\description{
\code{pl_summarize()} returns one row for each combination of grouping variables
(one difference with \code{dplyr::summarize()} is that \code{pl_summarize()} only
accepts grouped data). It will contain one column for each grouping variable
and one column for each of the summary statistics that you have specified.
}
\examples{
mtcars |>
  as_polars() |>
  group_by(cyl) |>
  pl_summarize(gear = mean(gear), gear2 = sd(gear))
}
