% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull.R
\name{pull.RPolarsDataFrame}
\alias{pull.RPolarsDataFrame}
\alias{pull.RPolarsLazyFrame}
\title{Extract a variable of a Data/LazyFrame}
\usage{
\method{pull}{RPolarsDataFrame}(.data, var, ...)

\method{pull}{RPolarsLazyFrame}(.data, var, ...)
}
\arguments{
\item{.data}{A Polars Data/LazyFrame}

\item{var}{A quoted or unquoted variable name}

\item{...}{Not used.}
}
\description{
This returns an R vector and not a Polars Series.
}
\examples{
\dontshow{if (require("dplyr", quietly = TRUE) && require("tidyr", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
pl_test <- polars::pl$DataFrame(iris)
pull(pl_test, Sepal.Length)
pull(pl_test, "Sepal.Length")
\dontshow{\}) # examplesIf}
}
