% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{as_polars_df}
\alias{as_polars_lf}
\alias{is_polars_lf}
\alias{is_polars_df}
\alias{is_polars_expr}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{polars}{\code{\link[polars]{as_polars_df}}, \code{\link[polars]{as_polars_lf}}, \code{\link[polars:check_polars]{is_polars_df}}, \code{\link[polars:check_polars]{is_polars_expr}}, \code{\link[polars:check_polars]{is_polars_lf}}}
}}

