% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{relig_income}
\alias{relig_income}
\title{Pew religion and income survey}
\format{
A dataset with variables:
* religion: name of religion
* `<$10k-Don\'t know/refused`: Number of respondents with income range in
column name
}
\source{
Downloaded from <https://www.pewresearch.org/religion/religious-landscape-study/>
(downloaded November 2009)
}
\usage{
relig_income
}
\description{
This dataset is reexported from `tidyr::relig_income`
}
\keyword{datasets}
