% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sink.R
\name{sink_ipc}
\alias{sink_ipc}
\title{Stream output to an IPC file}
\usage{
sink_ipc(
  .data,
  path,
  ...,
  compression = "zstd",
  compat_level = "newest",
  maintain_order = TRUE,
  type_coercion = TRUE,
  predicate_pushdown = TRUE,
  projection_pushdown = TRUE,
  simplify_expression = TRUE,
  slice_pushdown = TRUE,
  no_optimization = FALSE
)
}
\arguments{
\item{.data}{A Polars LazyFrame.}

\item{path}{Output file.}

\item{...}{Ignored.}

\item{compression}{\code{NULL} or a character of the compression method,
\code{"uncompressed"} or "lz4" or "zstd". \code{NULL} is equivalent to \code{"uncompressed"}.
Choose "zstd" for good compression performance. Choose "lz4"
for fast compression/decompression.}

\item{compat_level}{Determines the compatibility level when exporting Polars'
internal data structures. When specifying a new compatibility level, Polars
exports its internal data structures that might not be interpretable by other
Arrow implementations. The level can be specified as the name (e.g.,
\code{"newest"}) or as a scalar integer (currently, 0 or 1 is supported).
\itemize{
\item \code{"newest"} (default): Use the highest level, currently same as 1 (Low
compatibility).
\item \code{"oldest"}: Same as 0 (High compatibility).
}}

\item{maintain_order}{Whether maintain the order the data was processed
(default is \code{TRUE}). Setting this to \code{FALSE} will be slightly faster.}

\item{type_coercion}{Coerce types such that operations succeed and run on
minimal required memory (default is \code{TRUE}).}

\item{predicate_pushdown}{Applies filters as early as possible at scan level
(default is \code{TRUE}).}

\item{projection_pushdown}{Select only the columns that are needed at the
scan level (default is \code{TRUE}).}

\item{simplify_expression}{Various optimizations, such as constant folding
and replacing expensive operations with faster alternatives (default is
\code{TRUE}).}

\item{slice_pushdown}{Only load the required slice from the scan. Don't
materialize sliced outputs level. Don't materialize sliced outputs (default
is \code{TRUE}).}

\item{no_optimization}{Sets the following optimizations to \code{FALSE}:
\code{predicate_pushdown}, \code{projection_pushdown},  \code{slice_pushdown},
\code{simplify_expression}. Default is \code{FALSE}.}
}
\value{
The input LazyFrame.
}
\description{
This function allows to stream a LazyFrame that is larger than RAM directly
to an IPC file without collecting it in the R session, thus preventing
crashes because of too small memory.
}
