% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_ipc_polars}
\alias{write_ipc_polars}
\title{Export data to IPC file(s)}
\usage{
write_ipc_polars(
  .data,
  file,
  compression = "uncompressed",
  ...,
  compat_level = "newest",
  future
)
}
\arguments{
\item{.data}{A Polars DataFrame.}

\item{file}{File path to which the result should be written.}

\item{compression}{\code{NULL} or a character of the compression method,
\code{"uncompressed"} or "lz4" or "zstd". \code{NULL} is equivalent to \code{"uncompressed"}.
Choose "zstd" for good compression performance. Choose "lz4"
for fast compression/decompression.}

\item{...}{Ignored.}

\item{compat_level}{Determines the compatibility level when exporting Polars'
internal data structures. When specifying a new compatibility level, Polars
exports its internal data structures that might not be interpretable by other
Arrow implementations. The level can be specified as the name (e.g.,
\code{"newest"}) or as a scalar integer (currently, 0 or 1 is supported).
\itemize{
\item \code{"newest"} (default): Use the highest level, currently same as 1 (Low
compatibility).
\item \code{"oldest"}: Same as 0 (High compatibility).
}}

\item{future}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Deprecated, use
\code{compat_level} instead.}
}
\value{
The input DataFrame.
}
\description{
Export data to IPC file(s)
}
\details{
\subsection{Partitioned output}{

It is possible to export data to multiple files based on various parameters,
such as the values of some variables, or such that each file has a maximum
number of rows. See \code{\link[=partition_by]{partition_by()}} for more details.
}
}
