% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_ipc_polars}
\alias{write_ipc_polars}
\title{Export data to IPC file(s)}
\usage{
write_ipc_polars(
  .data,
  file,
  compression = "uncompressed",
  ...,
  future = FALSE
)
}
\arguments{
\item{.data}{A Polars DataFrame.}

\item{file}{File path to which the result should be written.}

\item{compression}{\code{NULL} or a character of the compression method,
\code{"uncompressed"} or "lz4" or "zstd". \code{NULL} is equivalent to \code{"uncompressed"}.
Choose "zstd" for good compression performance. Choose "lz4"
for fast compression/decompression.}

\item{...}{Ignored.}

\item{future}{Setting this to \code{TRUE} will write Polars' internal data
structures that might not be available by other Arrow implementations.}
}
\value{
The input DataFrame.
}
\description{
Export data to IPC file(s)
}
