% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_json_polars}
\alias{write_json_polars}
\title{Export data to JSON file(s)}
\usage{
write_json_polars(.data, file, ..., pretty = FALSE, row_oriented = FALSE)
}
\arguments{
\item{.data}{A Polars DataFrame.}

\item{file}{File path to which the result should be written.}

\item{...}{Ignored.}

\item{pretty}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Deprecated with no
replacement.}

\item{row_oriented}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Deprecated with no
replacement.}
}
\value{
The input DataFrame.
}
\description{
Export data to JSON file(s)
}
\examples{
\dontshow{if (requireNamespace("jsonlite")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
dest <- tempfile(fileext = ".json")
mtcars |>
  as_polars_df() |>
  write_json_polars(dest)

jsonlite::fromJSON(dest)
\dontshow{\}) # examplesIf}
}
