% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_ndjson_polars}
\alias{write_ndjson_polars}
\title{Export data to NDJSON file(s)}
\usage{
write_ndjson_polars(.data, file)
}
\arguments{
\item{.data}{A Polars DataFrame.}

\item{file}{File path to which the result should be written.}
}
\value{
The input DataFrame.
}
\description{
Export data to NDJSON file(s)
}
\examples{
\dontshow{if (requireNamespace("jsonlite")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
dest <- tempfile(fileext = ".ndjson")
mtcars |>
  as_polars_df() |>
  write_ndjson_polars(dest)

jsonlite::stream_in(file(dest), verbose = FALSE)
\dontshow{\}) # examplesIf}
}
